pyproject.toml
[project]
name = "retrieval-based-voice-conversion-webui"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "absl-py>=1.2.0",
    "audioread>=3.1.0",
    "av>=17.0.0",
    "colorama>=0.4.5",
    "cython>=3.2.4",
    "dfnstream-py>=0.2.1",
    "fairseq",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.88",
    "ffmpeg-python>=0.2.0",
    "ffmpy==0.3.1",
    "fsspec>=2022.11.0",
    "gradio==3.34.0",
    "gradio-client==0.2.7",
    "httpx>=0.28.1",
    "jinja2>=3.1.2",
    "joblib>=1.1.0",
    "json5>=0.13.0",
    "librosa==0.10.2",
    "llvmlite>=0.46.0",
    "markdown>=3.10.2",
    "matplotlib>=3.7.0",
    "matplotlib-inline>=0.1.3",
    "numba>=0.64.0",
    "numpy>=2.2.6",
    "onnxruntime>=1.24.3 ; sys_platform == 'darwin'",
    "onnxruntime-gpu>=1.24.3 ; sys_platform != 'darwin'",
    "pillow>=9.1.1",
    "praat-parselmouth>=0.4.2",
    "pyasn1>=0.4.8",
    "pyasn1-modules>=0.2.8",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "pyworld==0.3.2",
    "pyyaml>=6.0",
    "resampy>=0.4.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "soundfile>=0.12.1",
    "sympy>=1.11.1",
    "tabulate>=0.8.10",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torchcrepe==0.0.23",
    "torchfcpe>=0.0.4",
    "tornado>=6.1",
    "tqdm>=4.63.1",
    "uc-micro-py>=1.0.1",
    "uvicorn>=0.21.1",
    "websockets>=15.0.1",
    "werkzeug>=2.2.3",
]

[tool.uv.sources]
fairseq = { path = "./fairseq" }
